refactor(crawler): extract sitemap discovery into its own module (#442)
Move crawl_with_sitemap, crawl_with_sitemap_internal,
crawl_with_subpath_sitemaps and discover_sitemap_url out of the
discovery.rs God Object into application/crawler/sitemap_discovery.rs.

discovery.rs re-exports crawl_with_sitemap so the crawler facade and all
external callers (CLI, MCP, tests) resolve unchanged. Pure move, no
behavior change.
